sb/requirements.txt
wheel
Django==3.0.7
bs4
requests
pyaml
django-grappelli
Pillow
markdown
cryptography==3.1
social-auth-core==3.3.3
social-auth-app-django

#git+https://github.com/liqpay/sdk-python#egg=liqpay
django-mptt
psycopg2-binary
bs4
python-telegram-bot==12.0.0b1
wheel
celery==4.4.6
redis==3.2.0
django-tagging
django-crispy-forms
djangorestframework
drf-yasg